[expander] cross-phase persistent: allow more quote, quote-syntax
Previously, `quote` at the module level (i.e. on the rhs of a `define-values`
and not wrapped in a lambda, i.e. would run at module instantiation time) was
limited to a small subset of atomic data, and `quote-syntax` was forbidden to
appear anywhere in the module.

Loosen these restrictions, significantly. Specifically:

Say that simple data are: numbers, booleans, symbols, chars, keywords, null,
immutable strings, immutable bytes, immutable vectors with simple data for
values, immutable hasheqs with simple data for keys and values, and immutable
boxes with simple data for content. (These correspond roughly to things which
the standard reader can read. It excludes: extflonums, because their predicate
isn't in `#%kernel`; non-hasheq hashes, because hasheqs were previously allowed
as values in xpp modules but other hashes weren't; prefab structs, because I
was lazy.)

The restrictions for `quote` are then relaxed so that any simple data can be
quoted at the module level. (As before, arbitrary data can be quoted when
guarded by a lambda.)

For `quote-syntax`, the restriction to simple data is enforced throughout the
module. This is necessary because the stx's quoted by a quote-syntax are, in
essence, lifted to the module level.