make `#lang racket/kernel` entirely kernel
The catch before was that the reader submod used `module-reader`, which is
quite complex and decidedly not written in `#%kernel`. However, perhaps
unsurprising given `#lang racket/kernel`'s simplicity, only the simplest
features of module-reader were used, and it is quite straightforward to
simply inline the necessary components and rewrite them in `#%kernel`.

This change means that the `#lang` syntax can be used to write modules
which are entirely independent from everything other than "kernel.rkt"
itself.